# public/api-documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agenda API ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://0.0.0.0:8080",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/v1/agendas/adicionar": {
            "post": {
                "tags": [
                    "Agendas"
                ],
                "summary": "Adiioncar nova atividade na agenda.",
                "operationId": "adicionarAtividade",
                "requestBody": {
                    "description": "Dados da requisição.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data_inicio",
                                    "data_prazo",
                                    "status",
                                    "titulo",
                                    "descricao",
                                    "tipo_atividade_id"
                                ],
                                "properties": {
                                    "data_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-30"
                                    },
                                    "data_prazo": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-11-10"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "titulo": {
                                        "type": "string",
                                        "example": "Minha Atividade"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição da atividade"
                                    },
                                    "tipo_atividade_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Atividade adicionada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Atividade adicionada com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflito de domínio.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Conflito de domínio."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Houve um erro não esperado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/v1/agendas/atualizar": {
            "put": {
                "tags": [
                    "Agendas"
                ],
                "summary": "Atualiza uma atividade na agenda.",
                "operationId": "atualizarAtividade",
                "requestBody": {
                    "description": "Dados da atividade",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data_inicio",
                                    "data_prazo",
                                    "status",
                                    "titulo",
                                    "descricao",
                                    "tipo_atividade_id"
                                ],
                                "properties": {
                                    "data_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-30"
                                    },
                                    "data_prazo": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-11-10"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "titulo": {
                                        "type": "string",
                                        "example": "Minha Atividade Atualizada"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição atualizada da atividade"
                                    },
                                    "tipo_atividade_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atividade atualizada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Atividade atualizada com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflito de domínio.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Conflito de domínio."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Houve um erro não esperado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agendas/filtrar": {
            "post": {
                "tags": [
                    "Agendas"
                ],
                "summary": "Filtra as atividades do usuário.",
                "description": "Filtra as atividades do usuário com base nos parâmetros fornecidos.",
                "operationId": "filtrarAtividades",
                "requestBody": {
                    "description": "Parâmetros de filtragem",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "data_prazo": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lista de atividades filtradas."
                    },
                    "422": {
                        "description": "Erro de validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erros de validação encontrados."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "data_inicio": [
                                                    "Data obrigatória"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/agendas/remover/{agendaId}": {
            "delete": {
                "tags": [
                    "Agendas"
                ],
                "summary": "Remove uma atividade da agenda.",
                "operationId": "removerAtividade",
                "parameters": [
                    {
                        "name": "agendaId",
                        "in": "path",
                        "description": "ID da atividade a ser removida",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Atividade removida com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Atividade removida com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Atividade não encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Atividade não encontrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Houve um erro não esperado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/usuarios/cadastrar": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Cadastrar um novo usuário",
                "operationId": "2c4d732214932df996e20b32213a2d5c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário criado com sucesso"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Token de acesso obtido na autenticação",
                "name": "Autorização",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}